Add criterion-rs benches
