+++ exonum/Cargo.toml
@@ -62,6 +62,13 @@ pretty_assertions = "0.5.0"
 tempdir = "0.3.5"
 lazy_static = "1.0.0"
 proptest = "0.6.0"
+criterion = "0.2"
+num = "0.1.42"
+
+[[bench]]
+name = "criterion"
+harness = false
+path = "benches/criterion/lib.rs"
 
 [features]
 default = []
